feat: rename instrumentation module
