+++ executions/graphql-kotlin-dataloader-instrumentation/src/main/kotlin/com/expediagroup/graphql/dataloader/instrumentation/exceptions/MissingDataLoaderRegistryException.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2022 Expedia, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.expediagroup.graphql.dataloader.instrumentation.exceptions
+
+import graphql.GraphQLContext
+import org.dataloader.DataLoaderRegistry
+
+/**
+ * Thrown when an instance of [DataLoaderRegistry] does not exists in the [GraphQLContext].
+ */
+class MissingDataLoaderRegistryException : RuntimeException("DataLoaderRegistry instance not found in the GraphQLContext")
