+++ executions/graphql-kotlin-dataloader-instrumentation/src/test/kotlin/com/expediagroup/graphql/dataloader/instrumentation/datafetcher/NasaService.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2022 Expedia, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.expediagroup.graphql.dataloader.instrumentation.datafetcher
+
+import graphql.schema.DataFetchingEnvironment
+import java.util.concurrent.CompletableFuture
+
+class NasaService(
+    private val astronautService: AstronautService,
+    private val missionService: MissionService
+) {
+    fun getAstronaut(environment: DataFetchingEnvironment): CompletableFuture<Astronaut> =
+        astronautService
+            .getAstronaut(
+                AstronautServiceRequest(environment.getArgument<String>("id").toInt()),
+                environment
+            )
+
+    fun getMission(environment: DataFetchingEnvironment): CompletableFuture<Mission> =
+        missionService
+            .getMission(
+                MissionServiceRequest(environment.getArgument<String>("id").toInt()),
+                environment
+            )
+}
