+++ executions/graphql-kotlin-dataloader-instrumentation/src/test/kotlin/com/expediagroup/graphql/dataloader/instrumentation/fixture/domain/Astronaut.kt
@@ -1,0 +1,3 @@
+package com.expediagroup.graphql.dataloader.instrumentation.fixture.domain
+
+data class Astronaut(val id: Int, val name: String)
