+++ executions/graphql-kotlin-dataloader-instrumentation/src/test/kotlin/com/expediagroup/graphql/dataloader/instrumentation/fixture/domain/Mission.kt
@@ -1,0 +1,3 @@
+package com.expediagroup.graphql.dataloader.instrumentation.fixture.domain
+
+data class Mission(val id: Int, val designation: String, val crew: List<Int>)
