--- executions/graphql-kotlin-transaction-loader-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/NoOpExecutionStrategyInstrumentationContext.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2022 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.transactionbatcher.instrumentation
-
-import graphql.ExecutionResult
-import graphql.execution.instrumentation.ExecutionStrategyInstrumentationContext
-import java.util.concurrent.CompletableFuture
-
-/**
- * Simple NoOp [ExecutionStrategyInstrumentationContext] implementation
- */
-object NoOpExecutionStrategyInstrumentationContext : ExecutionStrategyInstrumentationContext {
-    override fun onDispatched(result: CompletableFuture<ExecutionResult>) {
-    }
-    override fun onCompleted(result: ExecutionResult, t: Throwable) {
-    }
-}
