--- executions/graphql-kotlin-transaction-loader-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/extensions/FieldValueInfosExtensions.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2022 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.transactionbatcher.instrumentation.extensions
-
-import graphql.execution.FieldValueInfo
-
-internal fun List<FieldValueInfo>.getExpectedStrategyCalls(): Int {
-    var count = 0
-    this.forEach { fieldValueInfo ->
-        when (fieldValueInfo.completeValueType) {
-            FieldValueInfo.CompleteValueType.OBJECT -> count++
-            FieldValueInfo.CompleteValueType.LIST -> count += fieldValueInfo.fieldValueInfos.getExpectedStrategyCalls()
-            else -> count += 0
-        }
-    }
-    return count
-}
