--- executions/graphql-kotlin-transaction-loader-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/TransactionLoaderLevelInstrumentation.kt
@@ -1,43 +1,0 @@
-/*
- * Copyright 2022 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level
-
-import com.expediagroup.graphql.dataloader.KotlinDataLoaderRegistry
-import com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution.AbstractExecutionLevelInstrumentation
-import com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution.ExecutionLevelInstrumentationContext
-import com.expediagroup.graphql.transactionbatcher.instrumentation.level.execution.ExecutionLevelInstrumentationParameters
-import com.expediagroup.graphql.transactionbatcher.instrumentation.level.state.Level
-import graphql.ExecutionInput
-import graphql.GraphQLContext
-
-/**
- * Once a certain [Level] is dispatched for all [ExecutionInput] sharing a [GraphQLContext] map
- * it will automatically dispatch a [KotlinDataLoaderRegistry] instance located in the GraphQLContext map.
- */
-class TransactionLoaderLevelInstrumentation : AbstractExecutionLevelInstrumentation() {
-    override fun calculateLevelState(
-        parameters: ExecutionLevelInstrumentationParameters
-    ): ExecutionLevelInstrumentationContext =
-        object : ExecutionLevelInstrumentationContext {
-            override fun onDispatched(level: Level, executions: List<ExecutionInput>) {
-                parameters
-                    .executionContext
-                    .graphQLContext.get<KotlinDataLoaderRegistry>(KotlinDataLoaderRegistry::class)
-                    ?.dispatchAll()
-            }
-        }
-}
