--- executions/graphql-kotlin-transaction-loader-instrumentation/src/main/kotlin/com/expediagroup/graphql/transactionbatcher/instrumentation/level/state/Level.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright 2022 Expedia, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.expediagroup.graphql.transactionbatcher.instrumentation.level.state
-
-/**
- * Represent a Level of a Document
- */
-data class Level(private val number: Int) {
-    /**
-     * calculate [Level] that is after this [Level]
-     * @return next [Level]
-     */
-    fun next(): Level = Level(number + 1)
-
-    /**
-     * calculate [Level] that is before this [Level]
-     * @return previous [Level]
-     */
-    fun previous(): Level = Level(number - 1)
-
-    /**
-     * calculate if this [Level] is the first
-     * @return whether or not this is the first [Level]
-     */
-    fun isFirst(): Boolean = number == 1
-}
